+++ examples/storage/storage/users/resources.py
@@ -3,13 +3,13 @@
 :Authors: cykooz
 :Date: 24.01.2020
 """
-from pyramid.security import ALL_PERMISSIONS, Allow
-from storage.authentication.models import UserModel
-from storage.root import StorageRoot
+from pyramid.authorization import ALL_PERMISSIONS, Allow
 
 from restfw.hal import HalResource
 from restfw.typing import PyramidRequest
 from restfw.utils import create_validation_error
+from storage.authentication.models import UserModel
+from storage.root import StorageRoot
 
 
 class User(HalResource):