+++ src/restfw/resources.py
@@ -145,7 +145,7 @@ class sub_resource_config(object):
         self.parent = parent
         self.predicates = predicates
         self.depth = predicates.pop('_depth', 0)
-        self.category = predicates.pop('_category', 'restfw')
+        self.category = predicates.pop('_category', 'pyramid')
 
     def register(self, scanner, name, wrapped):
         config = scanner.config