+++ src/restfw/views.py
@@ -78,7 +78,7 @@ class resource_view_config:
         self.resource_class = resource_class
         self.predicates = predicates
         self.depth = predicates.pop('_depth', 0)
-        self.category = predicates.pop('_category', 'restfw')
+        self.category = predicates.pop('_category', 'pyramid')
 
     def register(self, scanner, name, wrapped):
         config: Configurator = scanner.config